+++ packages/rejourney/android/src/main/kotlin/com/rejourney/RejourneySdkInfo.kt
@@ -2,5 +2,5 @@ package com.rejourney
 
 object RejourneySdkInfo {
     @JvmStatic
-    var sdkVersion: String = "0.1.1"
+    var sdkVersion: String = "0.2.0"
 }